feat(game): display active pressing buttons
This involves designing and creating a GameState class to hold the game state.
I think that it's better for Game and Player to hold purely immutable data
(state-independent) and have all things that change in a separate state object.

So there's some refactoring going on here, and that's why this commit is a bit
huge. The Lucidchart diagram has been updated to match.