fix(sampling-master): manually cut off the sound
Seems like blink-based browser, when told to play a portion of an audio buffer,
it will stop a little bit too late (causing click sounds).

May suspection is that it waits for the next buffer loop before sound actually
stops.  My own experiements show that this problem is gone when you pad your
sounds with a silence of 512 samples (assuming 44100 hz samplerate)